Move Connection into src.
